DB SChema Changes

+++ Midas/Database/GreenBillsDb/Schema/Tables/Abbreviation.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[Abbreviation]
 (
-	[AbbreviationId] bigint identity(1,1) NOT NULL PRIMARY KEY,
+	[AbbreviationId] INT identity(1,1) NOT NULL PRIMARY KEY,
 	[Name] nvarchar(25) not null,
 	[Description] nvarchar(100)
 )
